Created classify_network_api
A mid-level class that contains common network API code
