+++ htmresearch/frameworks/nlp/classify_htm.py
@@ -19,28 +19,25 @@
 # http://numenta.org/licenses/
 # ----------------------------------------------------------------------
 
-import numpy
-import operator
 import os
 
 from tabulate import tabulate
 
-from nupic.data.file_record_stream import FileRecordStream
 from nupic.engine import Network
 
 from htmresearch.frameworks.classification.classification_network import (
   configureNetwork)
 from htmresearch.encoders.cio_encoder import CioEncoder
-from htmresearch.frameworks.nlp.classification_model import ClassificationModel
 from htmresearch.support.network_text_data_generator import NetworkDataGenerator
+from htmresearch.frameworks.nlp.classify_network_api import (
+  ClassificationNetworkAPI
+)
 
 
-
-class ClassificationModelHTM(ClassificationModel):
+class ClassificationModelHTM(ClassificationNetworkAPI):
   """Classify text using generic network-API based models."""
 
   def __init__(self,
-               networkConfig,
                retinaScaling=1.0,
                retina="en_associative",
                apiKey=None,
@@ -56,19 +53,11 @@ class ClassificationModelHTM(ClassificationModel):
     """
     super(ClassificationModelHTM, self).__init__(**kwargs)
 
-    self.networkConfig = networkConfig
     self.retinaScaling = retinaScaling
     self.retina = retina
     self.apiKey = apiKey
     self.network = self.initModel()
     self._initializeRegionHelpers()
-
-
-  def getClassifier(self):
-    """
-    Returns the classifier for the model.
-    """
-    return self.classifierRegion.getSelf().getAlgorithmInstance()
 
 
   def initModel(self):
@@ -82,34 +71,6 @@ class ClassificationModelHTM(ClassificationModel):
 
     # This encoder specifies the LanguageSensor output width.
     return configureNetwork(None, self.networkConfig, encoder)
-
-
-  def _initializeRegionHelpers(self):
-    """
-    Set helper member variables once network has been initialized. This will
-    also be called from _deSerializeExtraData()
-    """
-    learningRegions = []
-    for region in self.network.regions.values():
-      spec = region.getSpec()
-      if spec.parameters.contains('learningMode'):
-        learningRegions.append(region)
-
-    # Always a sensor and classifier region.
-    self.sensorRegion = self.network.regions[
-      self.networkConfig["sensorRegionConfig"].get("regionName")]
-    self.classifierRegion = self.network.regions[
-      self.networkConfig["classifierRegionConfig"].get("regionName")]
-
-    # There is sometimes a TP region
-    self.tpRegion = None
-    if self.networkConfig.has_key("tpRegionConfig"):
-      self.tpRegion = self.network.regions[
-        self.networkConfig["tpRegionConfig"].get("regionName")]
-
-    self.learningRegions = learningRegions
-
-    self.network.enableProfiling()
 
 
   def tokenize(self, text):
@@ -127,20 +88,6 @@ class ClassificationModelHTM(ClassificationModel):
     for sentence in sentenceList:
       tokenList.extend(sentence.split(","))
     return tokenList
-
-
-  def reset(self):
-    """
-    Issue a reset signal to the model. The assumption is that a sequence has
-    just ended and a new sequence is about to begin.  The default behavior is
-    to do nothing - not all subclasses may re-implement this.
-    """
-    # TODO: Introduce a consistent reset method name.
-    for region in self.learningRegions:
-      if region.type == "py.TemporalPoolerRegion":
-        region.executeCommand(["reset"])
-      elif region.type == "py.TPRegion":
-        region.executeCommand(["resetSequenceStates"])
 
 
   def trainToken(self, token, labels, sampleId, reset=0):
@@ -229,100 +176,3 @@ class ClassificationModelHTM(ClassificationModel):
       sortedSampleIdList.append(partitionIdList[i])
 
     return categoryLikelihoods, sortedSampleIdList, sortedDistances
-
-
-  def printRegionOutputs(self):
-    """
-    Print the outputs of regions to console for debugging, depending on
-    verbosity level.
-    """
-
-    print "================== HTM Debugging output:"
-    print "Sensor output:",
-    print self.sensorRegion.getOutputData("dataOut").nonzero()
-    print "Sensor categoryOut:",
-    print self.sensorRegion.getOutputData("categoryOut")
-
-    if self.verbosity >= 3:
-      if self.tpRegion is not None:
-        print "TP region input:",
-        print self.tpRegion.getInputData("activeCells").nonzero()
-        print "TP region output:",
-        print self.tpRegion.getOutputData("mostActiveCells").nonzero()
-
-      print "Classifier bottomUpIn: ",
-      print self.classifierRegion.getInputData("bottomUpIn").nonzero()
-      print "Classifier categoryIn: ",
-      print self.classifierRegion.getInputData("categoryIn")[0:self.numLabels]
-
-    print "Classifier categoriesOut: ",
-    print self.classifierRegion.getOutputData("categoriesOut")[0:self.numLabels]
-    print "Classifier categoryProbabilitiesOut",
-    print self.classifierRegion.getOutputData("categoryProbabilitiesOut")[0:self.numLabels]
-
-
-  def dumpProfile(self):
-    """
-    Print region profiling information in a nice format.
-    """
-    print "Profiling information for {}".format(type(self).__name__)
-    totalTime = 0.000001
-    for region in self.network.regions.values():
-      timer = region.computeTimer
-      totalTime += timer.getElapsed()
-
-    profileInfo = []
-    for region in self.network.regions.values():
-      timer = region.computeTimer
-      profileInfo.append([region.name,
-                          timer.getStartCount(),
-                          timer.getElapsed(),
-                          100.0*timer.getElapsed()/totalTime])
-
-    profileInfo.append(["Total time", "", totalTime, "100.0"])
-    print tabulate(profileInfo, headers=["Region", "Count",
-                   "Elapsed", "Percent of total"],
-                   tablefmt = "grid")
-
-
-  def __getstate__(self):
-    """
-    Return serializable state.  This function will return a version of the
-    __dict__ with data that shouldn't be pickled stripped out. For example,
-    Network API instances are stripped out because they have their own
-    serialization mechanism.
-
-    See also: _serializeExtraData()
-    """
-    state = self.__dict__.copy()
-    # Remove member variables that we can't pickle
-    state.pop("network")
-    state.pop("sensorRegion")
-    state.pop("classifierRegion")
-    state.pop("tpRegion")
-    state.pop("learningRegions")
-
-    return state
-
-
-  def _serializeExtraData(self, extraDataDir):
-    """
-    Protected method that is called during serialization with an external
-    directory path. We override it here to save the Network API instance.
-
-    @param extraDataDir (string) Model's extra data directory path
-    """
-    self.network.save(os.path.join(extraDataDir, "network.nta"))
-
-
-  def _deSerializeExtraData(self, extraDataDir):
-    """
-    Protected method that is called during deserialization (after __setstate__)
-    with an external directory path. We override it here to load the Network API
-    instance.
-
-    @param extraDataDir (string) Model's extra data directory path
-    """
-    self.network = Network(os.path.join(extraDataDir, "network.nta"))
-    self._initializeRegionHelpers()
-    self.networkDataGen = NetworkDataGenerator()
